- добавлен отчет за месяц - исправление в форматах транзакций, для account id - исправлена возможность пустого названия счета

+++ src/main/java/ru/yandex/finance_tracker/controller/AccountController.java
@@ -10,7 +10,6 @@ import ru.yandex.finance_tracker.dto.input.AccountCreateRequest;
 import ru.yandex.finance_tracker.dto.output.AccountInfoDto;
 import ru.yandex.finance_tracker.security.dto.AuthInfo;
 import ru.yandex.finance_tracker.dto.output.TransactionInfoDto;
-import ru.yandex.finance_tracker.security.service.AuthenticationService;
 import ru.yandex.finance_tracker.service.AccountService;
 
 import java.util.List;
@@ -39,7 +38,7 @@ public class AccountController {
     @GetMapping("/{accountId}/transaction")
     @ResponseStatus(HttpStatus.OK)
     public List<TransactionInfoDto> getTransaction(
-            @PathVariable(name = "accountId") Integer accountId,
+            @PathVariable(name = "accountId") Long accountId,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "20") int size) {
         return accountService.getTransactionsByAccountId(accountId, page, size);
